2/scrape/requirements.txt
selenium
parsel
pyyaml
